main.c and makefile ported
